Add files via upload
Between 1999 and 2009, writing on the ground of Mozambique led us to search for roots, memories, silences and affections. They crystallized in travel drawings, traces made by a plurality of people to the flavor of the waves of the Indian Ocean. As the backdrop to Ilha de Moçambique, a meeting place for indigenous cultures and knowledge.